+++ tests/std.scm
@@ -342,3 +342,12 @@
                                             `&(:done #f :value ,(/ 1 i))))))))
           (t.is (iterator->array obj) #(1 1/2 1/3 1/4 1/5))
           (t.is (Array.from (iterator->array obj)) #(1 0.5 0.3333333333333333 0.25 0.2)))))
+
+(test "std: string interpolation"
+      (lambda (t)
+        (let ((x 10) (y 20))
+          (t.snapshot #"this is string \" ${(+ x y)} hello ${(repr "hello" #t)}
+                        world
+                         x
+                    x"))))
+